pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meraki-ap-exporter"
version = "0.1.1"
description = "Prometheus exporter focused on Cisco Meraki AP monitoring"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "prometheus-client>=0.21.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["main", "settings", "meraki_client", "exporter"]
